+++ dashboard/src/components/layout/MainNav/NavTree.tsx
@@ -489,9 +489,9 @@ export default function NavTree() {
                 }
               }}
               className={cn(
-                'flex h-8 w-full flex-row justify-start gap-1 bg-background px-1 text-foreground hover:bg-accent dark:hover:bg-muted',
+                'flex h-8 w-full flex-row justify-start gap-1 bg-background px-1 text-foreground hover:bg-accent',
                 {
-                  'bg-[#ebf3ff] hover:bg-[#ebf3ff] dark:bg-muted':
+                  'bg-[#ebf3ff] hover:bg-accent dark:bg-muted':
                     context.isFocused,
                 },
                 item.data.disabled && 'pointer-events-none opacity-50',