+++ dashboard/src/components/ui/v3/inline-code.tsx
@@ -5,11 +5,11 @@ export function InlineCode({
   children,
   className,
   ...props
-}: PropsWithChildren<{ className?: string }>) {
+}: PropsWithChildren<React.HTMLAttributes<HTMLElement>>) {
   return (
     <code
       className={cn(
-        'relative rounded bg-[#eaedf0] px-1 font-mono text-[11px] dark:bg-[#2f363d]',
+        'relative max-w-xs truncate rounded bg-[#eaedf0] px-1 font-mono text-[11px] dark:bg-[#2f363d]',
         className,
       )}
       {...props}