+++ dashboard/src/features/orgs/projects/database/dataGrid/components/DataBrowserGrid/DataBrowserGrid.tsx
@@ -15,17 +15,17 @@ import { normalizeDefaultValue } from '@/features/orgs/projects/database/dataGri
 import {
   POSTGRESQL_CHARACTER_TYPES,
   POSTGRESQL_DATE_TIME_TYPES,
-  POSTGRESQL_DECIMAL_TYPES,
-  POSTGRESQL_INTEGER_TYPES,
   POSTGRESQL_JSON_TYPES,
+  POSTGRESQL_NUMERIC_TYPES,
 } from '@/features/orgs/projects/database/dataGrid/utils/postgresqlConstants';
 import type { DataGridProps } from '@/features/orgs/projects/storage/dataGrid/components/DataGrid';
 import { DataGrid } from '@/features/orgs/projects/storage/dataGrid/components/DataGrid';
 import { DataGridBooleanCell } from '@/features/orgs/projects/storage/dataGrid/components/DataGridBooleanCell';
 import { DataGridDateCell } from '@/features/orgs/projects/storage/dataGrid/components/DataGridDateCell';
-import { DataGridDecimalCell } from '@/features/orgs/projects/storage/dataGrid/components/DataGridDecimalCell';
-import { DataGridIntegerCell } from '@/features/orgs/projects/storage/dataGrid/components/DataGridIntegerCell';
+import { DataGridNumericCell } from '@/features/orgs/projects/storage/dataGrid/components/DataGridNumericCell';
 import { DataGridTextCell } from '@/features/orgs/projects/storage/dataGrid/components/DataGridTextCell';
+import { isNotEmptyValue } from '@/lib/utils';
+
 import { useQueryClient } from '@tanstack/react-query';
 import { KeyRound } from 'lucide-react';
 import dynamic from 'next/dynamic';
@@ -68,6 +68,7 @@ export function createDataGridColumn(
     isEditable,
     type: 'text',
     specificType: column.full_data_type,
+    dataType: column.data_type,
     maxLength: column.character_maximum_length,
     Cell: DataGridTextCell,
     isPrimary: column.is_primary,
@@ -82,21 +83,13 @@ export function createDataGridColumn(
     foreignKeyRelation: column.foreign_key_relation,
   };
 
-  if (POSTGRESQL_INTEGER_TYPES.includes(column.data_type)) {
+  if (POSTGRESQL_NUMERIC_TYPES.includes(column.data_type)) {
     return {
       ...defaultColumnConfiguration,
       type: 'number',
+      isCopiable: true,
       width: 250,
-      Cell: DataGridIntegerCell,
-    };
-  }
-
-  if (POSTGRESQL_DECIMAL_TYPES.includes(column.data_type)) {
-    return {
-      ...defaultColumnConfiguration,
-      type: 'text',
-      width: 250,
-      Cell: DataGridDecimalCell,
+      Cell: DataGridNumericCell,
     };
   }
 
@@ -137,6 +130,7 @@ export function createDataGridColumn(
       ...defaultColumnConfiguration,
       type: 'date',
       width: 200,
+      isCopiable: true,
       Cell: DataGridDateCell,
     };
   }
@@ -166,8 +160,12 @@ export default function DataBrowserGrid({
 
   const { mutateAsync: updateRow } = useUpdateRecordWithToastMutation();
 
+  const sortByString = isNotEmptyValue(sortBy?.[0])
+    ? `${sortBy[0].id}.${sortBy[0].desc}`
+    : 'default-order';
+
   const { data, status, error, refetch } = useTableQuery(
-    [currentTablePath, limit, currentOffset, sortBy],
+    [currentTablePath, currentOffset, sortByString],
     {
       limit,
       offset: currentOffset * limit,
@@ -274,6 +272,8 @@ export default function DataBrowserGrid({
           // TODO: Create proper typings for data browser columns
           columns={memoizedColumns as unknown as DataBrowserGridColumn[]}
           onSubmit={refetch}
+          currentOffset={currentOffset}
+          sortByString={sortByString}
         />
       ),
     });