+++ dashboard/src/features/orgs/projects/database/dataGrid/utils/postgresqlConstants/postgresqlConstants.ts
@@ -19,7 +19,7 @@ export const POSTGRESQL_ERROR_CODES = {
  *
  * @docs https://www.postgresql.org/docs/current/datatype-numeric.html
  */
-export const POSTGRESQL_INTEGER_TYPES = [
+export const POSTGRESQL_NUMERIC_TYPES = [
   'smallint',
   'integer',
   'bigint',
@@ -27,16 +27,21 @@ export const POSTGRESQL_INTEGER_TYPES = [
   'serial',
   'bigserial',
   'oid',
+  'numeric',
+  'real',
+  'double precision',
 ];
-
-export const POSTGRESQL_DECIMAL_TYPES = ['numeric', 'real', 'double precision'];
 
 /**
  * Character data types in PostgreSQL.
  *
  * @docs https://www.postgresql.org/docs/current/datatype-character.html
  */
-export const POSTGRESQL_CHARACTER_TYPES = ['varchar', 'character', 'text'];
+export const POSTGRESQL_CHARACTER_TYPES = [
+  'character varying',
+  'character',
+  'text',
+];
 
 /**
  * JSON data types in PostgreSQL.